utils: Add cp library (refs EVEREST-1270)
Add a new library for CP related utilities. This library will be used
by the tarragon driver to provide CP related functionality.
Moved functions from the tarragon driver to the new library.
