+++ modules/CbTarragonDriver/tarragon/CMakeLists.txt
@@ -14,10 +14,16 @@ target_sources(tarragon
 )
 
 get_property(LINUX_HW_INCLUDE_DIR GLOBAL PROPERTY LINUX_HW_INCLUDE_DIR)
+get_property(CP_INCLUDE_DIR GLOBAL PROPERTY CP_INCLUDE_DIR)
 
 get_target_property(GENERATED_INCLUDE_DIR generate_cpp_files EVEREST_GENERATED_INCLUDE_DIR)
 
-target_include_directories(tarragon PRIVATE ${LINUX_HW_INCLUDE_DIR} ${GENERATED_INCLUDE_DIR})
+target_include_directories(tarragon
+    PRIVATE 
+        ${LINUX_HW_INCLUDE_DIR}
+        ${CP_INCLUDE_DIR}
+        ${GENERATED_INCLUDE_DIR}
+    )
 
 add_dependencies(tarragon generate_cpp_files)
