+++ modules/CbTarragonDriver/tarragon/CbTarragonPWM.hpp
@@ -41,11 +41,6 @@ public:
     void set_duty_cycle(double duty_cycle);
 
     /// @brief Check whether the current duty cycle is nominal.
-    //  @param duty_cycle The duty cycle to check.
-    /// @return True when configured duty cycle is >0 and <100% (nominal duty cycle),
-    static bool is_nominal_duty_cycle(const double& duty_cycle);
-
-    /// @brief Check whether the current duty cycle is nominal.
     /// @return True when configured duty cycle is >0 and <100% (nominal duty cycle),
     ///         false otherwise.
     bool is_nominal_duty_cycle() const;
