+++ utils/CMakeLists.txt
@@ -1,3 +1,4 @@
 set(UTILS_INCLUDE_DIR "${CMAKE_CURRENT_SOURCE_DIR}" PARENT_SCOPE)
 
 add_subdirectory(linux_hw)
+add_subdirectory(cp)
